add utils dir
