+++ src/app/thrift-services/damsel/index.ts
@@ -1,7 +1,6 @@
 export * from './operations';
 export * from './domain.service';
-export * from './get-thrift-instance';
-export * from './create-damsel-instance';
+export * from './utils/create-damsel-instance';
 export * from './damsel.module';
 export * from './routing-rules';
 export * from './services';
