+++ src/app/thrift-services/damsel/merchant-statistics.service.ts
@@ -4,10 +4,10 @@ import { map } from 'rxjs/operators';
 
 import { KeycloakTokenInfoService } from '../../keycloak-token-info.service';
 import { ThriftService } from '../thrift-service';
-import { createDamselInstance, damselInstanceToObject } from './create-damsel-instance';
 import { StatRequest, StatResponse } from './gen-model/merch_stat';
 import * as MerchantStatistics from './gen-nodejs/MerchantStatistics';
 import { StatRequest as ThriftStatRequest } from './gen-nodejs/merch_stat_types';
+import { createDamselInstance, damselInstanceToObject } from './utils/create-damsel-instance';
 
 @Injectable()
 export class MerchantStatisticsService extends ThriftService {
