+++ src/app/thrift-services/index.ts
@@ -1,6 +1,7 @@
 export * from './thrift-service';
 export * from './filters';
-export * from './thrift-instance';
+export * from './utils/thrift-instance';
+export * from './utils/get-thrift-instance';
 export * from './ank';
 export * from './file-storage';
 export * from './messages';
