+++ src/app/thrift-services/skipper/skipper-instance-utils.ts
@@ -1,5 +1,5 @@
 import metadata from '../../../assets/meta-skipper.json';
-import { createThriftInstanceUtils } from '../thrift-instance';
+import { createThriftInstanceUtils } from '../utils/thrift-instance';
 import * as base from './gen-nodejs/base_types';
 import * as chargeback from './gen-nodejs/chargeback_types';
 import * as skipper from './gen-nodejs/skipper_types';
