+++ src/app/thrift-services/utils/thrift-instance/namespace-type.ts
